* new version 2.19.9999
   * crash on importing certificates that do not have an email address
     associated to them, such as those of a Certificate Authority.

   * Disable saving new passwords to the password file. Implemented
     by Louis Raphael from dpslabs.com.

   * Panda IMAP does not decode correctly Korean text encoded in UTF-8.
     Reported by Chulho Yang.

+++ web/src/alpined.d/alpined.c
@@ -10844,11 +10844,11 @@ peLoadConfig(struct pine *pine_state)
 
     if(pine_state->VAR_RSHOPENTIMEO
 	&& ((rv = atoi(pine_state->VAR_RSHOPENTIMEO)) == 0 || rv > 4))
-      mail_parameters(NULL, SET_RSHTIMEOUT, (void *) rv);
+      mail_parameters(NULL, SET_RSHTIMEOUT, (void *) (long) rv);
 
     if(pine_state->VAR_SSHOPENTIMEO
 	&& ((rv = atoi(pine_state->VAR_SSHOPENTIMEO)) == 0 || rv > 4))
-      mail_parameters(NULL, SET_SSHTIMEOUT, (void *) rv);
+      mail_parameters(NULL, SET_SSHTIMEOUT, (void *) (long) rv);
 
     /*
      * Tell c-client not to be so aggressive about uid mappings
